Pick Up Feedback Sounds
As a player, I want to hear a celebratory chirp when I successfully use an object so I know what has happened.

As a player, I want to hear a low, discordant tone when I unsuccessfully use an object so I know what has happened and can try to correct my behavior.